Split data adapter functionality to different files
